Test CodeQL
content for CI must be longer

+++ core/packet.c
@@ -338,13 +338,13 @@ static int prv_send_new_block1(lwm2m_context_t * contextP, lwm2m_transaction_t *
 {
     lwm2m_transaction_t * next;
     // Done sending block
-    if (block_num * block_size >= previous->payload_len)
+    if ((size_t)block_num * block_size >= previous->payload_len)
         return 0;
 
     next = prv_create_next_block_transaction(previous, contextP->nextMID++);
     if (next == NULL) return COAP_500_INTERNAL_SERVER_ERROR;
 
-    size_t remaining_payload_length = next->payload_len - block_num * block_size;
+    size_t remaining_payload_length = next->payload_len - (size_t)block_num * block_size;
     uint8_t *new_block_start = next->payload + block_num * block_size;
 
     coap_set_header_block1(next->message, block_num, remaining_payload_length > block_size, block_size);
